+++ plugins/paper/src/main/java/com/playmonumenta/plugins/abilities/rogue/ByMyBlade.java
@@ -52,7 +52,7 @@ public class ByMyBlade extends Ability {
 	public void onDamage(DamageEvent event, LivingEntity enemy) {
 		if (event.getType() == DamageType.MELEE || event.getType() == DamageType.MELEE_ENCH || event.getType() == DamageType.MELEE_SKILL) {
 			int hasteAmplifier = getAbilityScore() == 1 ? BY_MY_BLADE_1_HASTE_AMPLIFIER : BY_MY_BLADE_2_HASTE_AMPLIFIER;
-			int extraDamage = mDamageBonus;
+			double extraDamage = mDamageBonus;
 
 			mPlugin.mPotionManager.addPotion(mPlayer, PotionID.ABILITY_SELF,
 			                                 new PotionEffect(PotionEffectType.FAST_DIGGING, BY_MY_BLADE_HASTE_DURATION, hasteAmplifier, false, true));
